fix: task link origin lookup
Ensure that we actually have the origin in the origin map before
attempting to dereference it. Since we do not hold the task link map
lock for longer than we need, the trade-off for performance is a
potential loss os sync with the data collected in earlier stages of the
unwind_tasks method. Therefore we do not make the assumption that the
lookup is always successful to prevent potential invalid memory
dereferencing.