fix: CPU clock ID retrieval on Linux
Trying to retrieve the CPU clock ID using an invalid pthread_t can cause
a SIGSEGV. We try to protect against this by first trying to copy a
chunk of memory from the address pointed by the thread ID (effectively a
pthread_t). If that fails we give up constructing the ThreadInfo object.