fix: track main thread only if missing
We make sure we don't accidentally mistaken a native thread for
the main thread when we don't find its native ID within the thread
tracking map. This can happen when attaching to a running process
using hypno, in which case we are likely to be missing the main
thread. There are cases where we can still get into the logic that
supports this case, but for native threads that are not the main
thread. When these threads die, they won't untrack themselves,
causing us to look up their clock ID on Linux and crash.