perf: interpreter state field read (#123)
The interpreter state is a large structure that spans across
multiple memory pages. Remote reads of this structure in a single
chunk suffer a performance penalty that can be avoided because we
only need a bunch of fields from it. It is therefore more
efficient to read the single fields than all the pages backing the
full structure.